+++ src/components/OHLCV/connect.js
@@ -1,7 +1,9 @@
 // @flow
 import { connect } from 'react-redux'
-import ohlcvModel, { updateTimeLine } from '../../store/models/ohlcv'
-import { saveDuration, saveTimeSpan } from '../../store/actions/ohlcv'
+import ohlcvModel, {
+  updateDuration,
+  updateTimeSpan,
+} from '../../store/models/ohlcv'
 
 import type { State } from '../../types'
 
@@ -10,9 +12,8 @@ export const mapStateToProps = (state: State) => {
 }
 
 export const mapDispatchToProps = {
-  updateTimeLine,
-  saveDuration,
-  saveTimeSpan
+  updateDuration,
+  updateTimeSpan,
 }
 
 export default connect(
