+++ src/store/models/ohlcv.js
@@ -1,7 +1,8 @@
 // @flow
+import { getTokenPairsDomain } from '../domains'
 import OHLCVModel from '../domains/ohlcv'
 import * as actionCreators from '../actions/ohlcv'
-import { timeSpans } from '../../components/OHLCV/SmallChart'
+import { timeSpans } from '../../components/OHLCV/OHLCV'
 import type { SendTimelineParams } from '../../types/ohlcv'
 import type { State, ThunkAction } from '../../types'
 
@@ -9,23 +10,62 @@ export default function getOHLCVModel(state: State) {
   return OHLCVModel(state.ohlcv)
 }
 
-export const updateTimeLine = ({ updateWRT }: SendTimelineParams): ThunkAction => {
+export const updateTimeLine = ({
+  updateWRT,
+}: SendTimelineParams): ThunkAction => {
   return async (dispatch, getState) => {
-    let currentDuration = getState().ohlcv.currentDuration.label
-    let currentTimeSpan = getState().ohlcv.currentTimeSpan.label
+    const currentDuration = getState().ohlcv.currentDuration.label
+    const currentTimeSpan = getState().ohlcv.currentTimeSpan.label
 
     if (updateWRT === 'timespan') {
-      let candles = byTimeSpan(currentDuration, currentTimeSpan)
-      if (candles > 40) {
+      const candles = byTimeSpan(currentDuration, currentTimeSpan)
+      if (candles > 40)
         return dispatch(actionCreators.saveNoOfCandles(parseInt(candles, 10)))
-      }
     } else {
-      let { candles, time } = byDuration(currentDuration)
-      if (time) {
-        dispatch(actionCreators.saveTimeSpan(timeSpans[time]))
-      }
+      const { candles, time } = byDuration(currentDuration)
+      if (time) dispatch(actionCreators.saveTimeSpan(timeSpans[time]))
       return dispatch(actionCreators.saveNoOfCandles(parseInt(candles, 10)))
     }
+  }
+}
+
+export const updateTimeSpan = (
+  currentTimeSpan: Object,
+  config: Object
+): ThunkAction => {
+  return (dispatch, getState, { socket }) => {
+    dispatch(actionCreators.saveTimeSpan(currentTimeSpan))
+    dispatch(updateTimeLine(config))
+
+    const state = getState()
+    const pairDomain = getTokenPairsDomain(state)
+    const currentPair = pairDomain.getCurrentPair()
+
+    socket.subscribeChart(
+      currentPair,
+      state.ohlcv.currentTimeSpan.label,
+      state.ohlcv.currentDuration.label
+    )
+  }
+}
+
+export const updateDuration = (
+  currentDuration: Object,
+  config: Object
+): ThunkAction => {
+  return (dispatch, getState, { socket }) => {
+    dispatch(actionCreators.saveDuration(currentDuration))
+    dispatch(updateTimeLine(config))
+
+    const state = getState()
+    const pairDomain = getTokenPairsDomain(state)
+    const currentPair = pairDomain.getCurrentPair()
+
+    socket.subscribeChart(
+      currentPair,
+      state.ohlcv.currentTimeSpan.label,
+      state.ohlcv.currentDuration.label
+    )
   }
 }
